# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ERB System API Documentation",
        "description": "Tài liệu đặc tả và kiểm thử RESTful API - Hệ thống quản trị ERB",
        "contact": {
            "email": "admin@erb.vn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/warehouses": {
            "get": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Lấy danh sách kho",
                "description": "Danh sách kho hỗ trợ tìm kiếm theo mã/tên, lọc theo công ty, chi nhánh, loại kho, kho cha, trạng thái và phân trang.",
                "operationId": "13113e11016af8a5ac9717747fff191c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số bản ghi/trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm theo mã hoặc tên kho",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company_id",
                        "in": "query",
                        "description": "Lọc theo công ty",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "branch_id",
                        "in": "query",
                        "description": "Lọc theo chi nhánh",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "warehouse_type",
                        "in": "query",
                        "description": "Loại kho (1: Kho tổng, 2: Kho con)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2
                            ]
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Lọc theo kho cha",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái (1: Hoạt động, 0: Ngừng hoạt động)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy danh sách thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "company_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "branch_id": {
                                                                "type": "integer",
                                                                "example": 1,
                                                                "nullable": true
                                                            },
                                                            "parent_id": {
                                                                "type": "integer",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "code": {
                                                                "type": "string",
                                                                "example": "WH001"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Kho Tổng HCM"
                                                            },
                                                            "warehouse_type": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "address": {
                                                                "type": "string",
                                                                "example": "123 Nguyễn Huệ, Quận 1"
                                                            },
                                                            "manager_id": {
                                                                "type": "integer",
                                                                "example": 1,
                                                                "nullable": true
                                                            },
                                                            "status": {
                                                                "type": "integer",
                                                                "example": 1
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 4
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Chưa xác thực tài khoản"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Tạo kho mới",
                "description": "Tạo kho tổng hoặc kho con. Mã kho (code) là duy nhất toàn hệ thống.",
                "operationId": "8c029d4e1b5dd0fe9c7d3cbef04628a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "company_id",
                                    "code",
                                    "name",
                                    "warehouse_type"
                                ],
                                "properties": {
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "parent_id": {
                                        "description": "Kho cha (chỉ với kho con)",
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "code": {
                                        "type": "string",
                                        "example": "WH010"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Kho Quận 7"
                                    },
                                    "warehouse_type": {
                                        "type": "integer",
                                        "enum": [
                                            1,
                                            2
                                        ],
                                        "example": 2
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Quận 7, TP.HCM",
                                        "nullable": true
                                    },
                                    "manager_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo kho thành công"
                    },
                    "400": {
                        "description": "Dữ liệu không hợp lệ"
                    },
                    "401": {
                        "description": "Chưa xác thực tài khoản"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/warehouses/{id}/disable": {
            "patch": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Ngừng hoạt động kho",
                "description": "Đặt trạng thái kho = 0 (Ngừng hoạt động). Kho ngừng hoạt động bị chặn mọi nghiệp vụ nhập/xuất/điều chuyển.",
                "operationId": "fdd7aa5ff5eade3bab4a0155cb2fca9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã ngừng hoạt động"
                    },
                    "404": {
                        "description": "Không tìm thấy kho"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/warehouses/{id}/enable": {
            "patch": {
                "tags": [
                    "Warehouses"
                ],
                "summary": "Kích hoạt lại kho",
                "description": "Đặt trạng thái kho = 1 (Hoạt động).",
                "operationId": "30701bc5cbd7e3238aa87da0e5749306",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã kích hoạt"
                    },
                    "404": {
                        "description": "Không tìm thấy kho"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Nhập JWT Access Token (dạng: Bearer <token>) để thực hiện các yêu cầu cần xác thực.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}